standardize data and introduce case insensitivity
sorry this looks like fixtures. But let caches values. So we were
getting cached values from previous tests for regexp. It worked in
minitest but was filing for rspecs. So going with this solution for now.

Maybe the above was a red herring, but introduced case sensitivity
so the way various databases work would pass tests.

I'm fully expecting case_sensitivity? to become quite complex as
mac pg and RHEL (sensitive) is different from travis/ubuntu pg (not)