Discover db config
Got dataset nuances out of specs and into the classes.
Detecting these nuances

Future: make ruby classes match db behavior

Changed databases to be more C/POSIX like.
So we don't have to deal with a < A and other locale nuances
PG mac and PG ubuntu differ, and don't want to detect all those issues

In the future, may support locales w/ comparison, but need a c extension
to support and the poor implementation on the mac makes punting the best
bet.